♻️ Move Pf2eAlignmentValue into JsonTextReplacement
